+++ src/EcMaster.h
@@ -50,7 +50,7 @@ public:
      * \param slaveInfo if true prints on EtherCATsoemInfo.txt the slave information.
      *    
      */
-    EcMaster(int cycleTime = 1000000, bool useDC = false, bool slaveInfo = false);
+    EcMaster(unsigned long cycleTime = 1000000, bool useDC = false, bool slaveInfo = false);
 
     /**
      *  \brief Destructor
@@ -115,7 +115,7 @@ private:
     std::string ethPort;
     char * ecPort;
     char m_IOmap[4096];
-    int m_cycleTime;	//the periodicity of ethercatLoop ("PDOs period")
+    unsigned long m_cycleTime;	//the periodicity of ethercatLoop ("PDOs period")
     bool m_useDC;
     std::vector<EcSlave*> m_drivers;
     int* offSetOutput;
